refactor(frontend): extract generic entity detail loading from IP view
Splits IP view into a generic template usable for any entity detail (data
loading, history selection, URL synchronization,...) and IP-specific charts.

The template will be reused for IP subnet view.
